+++ src/components/Table/tests/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Table from '../index';
-import { shallow } from 'enzyme';
+import { shallow, render } from 'enzyme';
 
 describe('components/Table', () => {
   it('should render as a table', () => {
@@ -24,6 +24,20 @@ describe('components/Table', () => {
     it('should render with default `density` prop value as OUI table class modifier', () => {
       const component = shallow(<Table></Table>);
       expect(component.hasClass('oui-table--loose')).toBe(false);
+    });
+  });
+
+  describe('with `tableLayoutAlgorithm` prop supplied', () => {
+    it('should render with provided `tableLayoutAlgorithm` prop value as inline style', () => {
+      const component = render(<Table tableLayoutAlgorithm='auto'></Table>);
+      expect(component.children().attr('style')).toContain('table-layout:auto');
+    });
+  });
+
+  describe('with no `tableLayoutAlgorithm` prop supplied', () => {
+    it('should render with default `tableLayoutAlgorithm` prop value as inline style', () => {
+      const component = render(<Table></Table>);
+      expect(component.children().attr('style')).toContain('table-layout:fixed');
     });
   });